+++ app/Http/Controllers/CompaniesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Company;
 use App\Http\Requests\CompaniesRequest as Request;
+use Illuminate\Support\Facades\Auth;
 
 class CompaniesController extends Controller
 {
@@ -14,8 +15,12 @@ class CompaniesController extends Controller
 
     public function store(Request $request)
     {
-        $company = Company::create($request->all());
-        return $company;
+        $data = $request->all();
+        $data['user_id'] = Auth::user()->id;
+        if (Auth::user()->type == "company" || Auth::user()->type == "admin") {
+            return Company::create($data);
+        }
+        return "Usuário não autorizado a cadastrar empresas";
     }
 
     public function show(Company $company)
@@ -26,15 +31,29 @@ class CompaniesController extends Controller
 
     public function update(Request $request, Company $company)
     {
-        $this->authorize('update', $company);
-        $company->update($request->all());
-        return $company;
+        if (Auth::user()->type == "client") {
+            $this->authorize('update', $company);
+            $company->update($request->all());
+            return $company;
+        }
+        if (Auth::user()->type == "admin") {
+            $company->update($request->all());
+            return $company;
+        }
+        return "Não está autorizado a editar esta empresa";
     }
 
     public function destroy(Company $company)
     {
-        $this->authorize('delete', $company);
-        $company->delete();
-        return $company;
+        if (Auth::user()->type == "client") {
+            $this->authorize('delete', $company);
+            $company->delete();
+            return $company;
+        }
+        if (Auth::user()->type == "admin") {
+            $company->delete();
+            return $company;
+        }
+        return "Não está autorizado a excluir esta empresa";
     }
 }
